Add confirmation for delete all logs

+++ al-settings.php
@@ -14,13 +14,13 @@
             <div id="icon-options-general" class="icon32"><br /></div>
 
             <h1>Action Logger settings page</h1>
-    
+
             <?php ActionLogger::al_show_admin_notices(); ?>
 
             <div id="action-logger" class="">
 
                 <?php echo al_admin_menu(); ?>
-    
+
                 <?php echo al_check_php_version(); ?>
 
                 <form name="settings-form" action="" method="post">
@@ -60,12 +60,11 @@
                 </div>
 
                 <h2><?php esc_html_e( 'Nuke \'em all', 'action-logger' ); ?></h2>
-                <p><?php esc_html_e( 'Delete all items. Watch out, there\'s no confirmation. Delete = delete !', 'action-logger' ); ?></p>
+                <p><?php // esc_html_e( 'Delete all items. Watch out, there\'s no confirmation. Delete = delete !', 'action-logger' ); ?></p>
                 <form name="delete-logs" action="" method="post">
                     <input name="delete_all_logs_nonce" type="hidden" value="<?php echo wp_create_nonce( 'delete-all-logs-nonce' ); ?>" />
                     <label for="delete-all-checkbox" class="screen-reader-text"><?php esc_html_e( 'Delete', 'action-logger' ); ?></label>
-                    <input name="delete_all_checkbox" id="delete-all-checkbox" type="checkbox" value="1" /> &nbsp;&nbsp;
-                    <input name="delete_all" type="submit" class="admin-button admin-button-small" value="<?php esc_html_e( 'Delete all', 'action-logger' ); ?>" />
+                    <input name="delete_all" type="submit" class="admin-button admin-button-small" onclick="return confirm( 'Are you sure you want to delete all logs ?' )" value="<?php esc_html_e( 'Delete all logs', 'action-logger' ); ?>" />
                 </form>
 
             </div><!-- end #action-logger -->
